Split renderer/sprite/spritesystem unit tests
